+++ source/renderer/app/actions/staking-actions.js
@@ -4,5 +4,5 @@ import Action from './lib/Action';
 // ======= STAKING ACTIONS =======
 
 export default class StakingActions {
-  goToStakingInfo: Action<any> = new Action();
+  goToStakingPage: Action<any> = new Action();
 }